refactor: simplify secret scanner partition coverage
